requirements.txt
streamlit
pandas
numpy
scikit-learn
pyyaml
nltk
sentence-transformers
faiss-cpu
